pega-mcp/requirements.txt
# MCP Server Dependencies
fastmcp>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# ADK Agent Dependencies
# google-adk>=1.2.0
